+++ src/App/components/Card/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const Card = props => {
+    return (
+        <section className="card">
+            asd
+        </section>
+    );
+};
+
+Card.propTypes = {};
+
+export default Card;